♻️ rename internal monitoring files -> telemetry
