+++ packages/core/src/domain/report/reportObservable.ts
@@ -1,7 +1,7 @@
 import { toStackTraceString } from '../../tools/error'
 import { mergeObservables, Observable } from '../../tools/observable'
 import { DOM_EVENT, includes, addEventListener, safeTruncate } from '../../tools/utils'
-import { monitor } from '../internalMonitoring'
+import { monitor } from '../telemetry'
 import type { Report, BrowserWindow, ReportType } from './browser.types'
 
 export const RawReportType = {
