+++ packages/core/src/domain/telemetry/telemetryEvent.types.ts
