+++ packages/core/src/tools/error.ts
@@ -1,4 +1,4 @@
-import { callMonitored } from '../domain/internalMonitoring'
+import { callMonitored } from '../domain/telemetry'
 import type { StackTrace } from '../domain/tracekit'
 import { computeStackTrace } from '../domain/tracekit'
 import type { ClocksState } from './timeUtils'
