+++ packages/core/src/tools/utils.ts
@@ -1,4 +1,4 @@
-import { monitor } from '../domain/internalMonitoring'
+import { monitor } from '../domain/telemetry'
 
 export const ONE_SECOND = 1000
 export const ONE_MINUTE = 60 * ONE_SECOND
