+++ test/e2e/scenario/telemetry.scenario.ts
